internship-email-extension/backend/requirements.txt
flask
flask-cors
python-dotenv
openai>=1.12.0
agents
